Move property input types to dedicated file
- Move PropertyOptionInput, PropertyFieldAttrsInput, PropertyFieldInput from graphql_playbook.go to graphql_property.go
- Better organization and separation of concerns
